refactor: centralize date helpers
