+++ choir-app-backend/src/controllers/availability.controller.js
@@ -1,20 +1,6 @@
 const db = require('../models');
 const { Op } = db.Sequelize;
-
-function datesForRule(year, month, rule) {
-    const dates = [];
-    const d = new Date(Date.UTC(year, month - 1, 1));
-    while (d.getUTCMonth() === month - 1) {
-        if (d.getUTCDay() === rule.dayOfWeek) {
-            const week = Math.floor((d.getUTCDate() - 1) / 7) + 1;
-            if (!Array.isArray(rule.weeks) || rule.weeks.length === 0 || rule.weeks.includes(week)) {
-                dates.push(new Date(Date.UTC(d.getUTCFullYear(), d.getUTCMonth(), d.getUTCDate())));
-            }
-        }
-        d.setUTCDate(d.getUTCDate() + 1);
-    }
-    return dates;
-}
+const { datesForRule, isoDateString } = require('../utils/date.utils');
 
 exports.findByMonth = async (req, res) => {
     const { year, month } = req.params;
@@ -23,7 +9,7 @@ exports.findByMonth = async (req, res) => {
         const dateSet = new Set();
         for (const rule of rules) {
             for (const d of datesForRule(year, month, rule)) {
-                dateSet.add(d.toISOString().split('T')[0]);
+                dateSet.add(isoDateString(d));
             }
         }
         const dates = Array.from(dateSet).sort();
